On branch master Your branch is up to date with 'origin/master'.
Changes to be committed:
	modified:   lib/app/data/models/users_model.dart
	modified:   lib/app/screens/home/petugas/request_pengambilan/detail_request.dart
	modified:   lib/app/screens/home/petugas/request_pengambilan/request_pengambilan_controller.dart
	modified:   lib/app/screens/home/profil/profil.dart
	modified:   lib/app/screens/home/profil/profil_controller.dart
	modified:   lib/app/screens/home/warga/kirim/kirim.dart
	modified:   lib/app/screens/home/warga/tukar/tukar.dart
	modified:   lib/app/screens/home/warga/tukar/tukar_controller.dart
	modified:   lib/service/authentication.dart
	modified:   lib/service/workmanager.dart

+++ lib/service/authentication.dart
@@ -169,7 +169,7 @@ class AuthController extends GetxController {
             rt: currUserData["rt"],
             rw: currUserData["rw"],
             wa: currUserData["no_wa"],
-            poin: currUserData['poin'],
+            poin: currUserData['poin'] as int,
             fotoProfil: currUserData["foto_profil"],
             rool: currUserData["rool"],
             password: currUserData["password"]));
